style: remove page-pdg class from auth pages

+++ server/db/product.js
@@ -10,7 +10,7 @@ const Product = db.define('product', {
     type: Sequelize.STRING,
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.DECIMAL(20, 2),
     defaultValue: 0.0,
   },
   quantity: {
